src/authapi/requirements.txt
Django==4.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
djangorestframework==3.15.0
PyJWT==2.8.0
pyotp==2.9.0
pillow==10.2.0